feat: encoder发送更新

+++ Embedded/Encoder/User/Encoder/encoder.h
@@ -30,7 +30,7 @@ typedef struct _EncoderInstance_s {
     bool is_auto_refresh;
     int refresh_time;
 
-    int angle;  //解码角度值
+    double angle;  //解码角度值
 } EncoderInstance_s;
